# backend-java/demo/src/main/java/com/example/demo/dto/RegisterRequest.java
package com.example.demo.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class RegisterRequest {
    @NotBlank(message = "用户名不能为空")
    private String username;

    private String email;

    @NotBlank(message = "密码不能为空")
    private String password;
}